moved highlight methods from stateManager to board
