+++ src/board/stateManager.js
@@ -67,7 +67,7 @@ export default class ChessStateManager {
         return col; // return the index of the column
     }
 
-    
+
     ///////////////////////
     /// OTHER UTILITIES /// 
     ///////////////////////
@@ -131,48 +131,48 @@ export default class ChessStateManager {
         return false;
     }
 
-    // MOVE TO BOARD
-    highlightMoves(currPos) {
-        let piece = this.state[currPos].getPiece();
-        piece.generateMoves(currPos);
+    // // MOVE TO BOARD --- READY TO MOVE
+    // highlightMoves(currPos) {
+    //     let piece = this.manager.getState()[currPos].getPiece();
+    //     piece.generateMoves(currPos);
 
-        let div = this.getDivFromIndex(currPos);
-        div.className = div.className + " highTile";
-        this.highlighted.push(div);
+    //     let div = this.manager.getDivFromIndex(currPos);
+    //     div.className = div.className + " highTile";
+    //     this.highlighted.push(div);
 
-        let moves = piece.getValidMoves();
-        moves = moves.concat(piece.getValidTakes());
+    //     let moves = piece.getValidMoves();
+    //     moves = moves.concat(piece.getValidTakes());
 
-        let takes = piece.getValidTakes();
+    //     let takes = piece.getValidTakes();
 
-        // moves and takes
-        for (let i = 0; i < moves.length; i++) {
-            div = this.getDivFromIndex(moves[i]);
-            if (div != null) {
-                this.highlighted.push(div);
-                div.className = div.className + " highTile";
-            }
-        }
+    //     // moves and takes
+    //     for (let i = 0; i < moves.length; i++) {
+    //         div = this.manager.getDivFromIndex(moves[i]);
+    //         if (div != null) {
+    //             this.highlighted.push(div);
+    //             div.className = div.className + " highTile";
+    //         }
+    //     }
 
-        // takes only
-        for (let i = 0; i < takes.length; i++) {
-            div = this.getDivFromIndex(takes[i]);
-            if (div != null) {
-                this.highlighted.push(div);
-                div.className = div.className + " underAttack";
-            }
-        }
-        //this.board.update(this.state);
+    //     // takes only
+    //     for (let i = 0; i < takes.length; i++) {
+    //         div = this.manager.getDivFromIndex(takes[i]);
+    //         if (div != null) {
+    //             this.highlighted.push(div);
+    //             div.className = div.className + " underAttack";
+    //         }
+    //     }
+    //     //this.board.update(this.state);
 
-    }
-    // MOVE TO BOARD
-    unhighlight() {
-        for (let i = 0; i < this.highlighted.length; i++) {
-            let div = this.highlighted[i];
-            div.className = div.className.replace(' highTile', '');
-            div.className = div.className.replace(' underAttack', '');
-        }
-    }
+    // }
+    // // MOVE TO BOARD
+    // unhighlight() {
+    //     for (let i = 0; i < this.highlighted.length; i++) {
+    //         let div = this.highlighted[i];
+    //         div.className = div.className.replace(' highTile', '');
+    //         div.className = div.className.replace(' underAttack', '');
+    //     }
+    // }
 
     // move: alter the game state and display the message
     move(currPos, newPos, isTake, takeName) {
